sync with wingtk/gvsbuild

+++ gvsbuild/utils/utils.py
@@ -1,6 +1,4 @@
-#  Copyright (C) 2016 - Yevgen Muntyan
-#  Copyright (C) 2016 - Ignacio Casal Quinteiro
-#  Copyright (C) 2016 - Arnavion
+#  Copyright (C) 2016 The Gvsbuild Authors
 #
 #  This program is free software; you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
@@ -16,54 +14,60 @@
 #  along with this program; if not, see <http://www.gnu.org/licenses/>.
 
 import os
+import re
+import shutil
 import stat
 import time
-import shutil
-import re
 
-from .simple_ui import print_debug
+from .simple_ui import log
+
 
 def convert_to_msys(path):
     path = path
-    if path[1] != ':':
-        raise Exception('oops')
-    path = '/' + path[0] + path[2:].replace('\\', '/')
-    return path
+    if path[1] != ":":
+        raise NotADirectoryError("Path doesn't contain a drive letter like C:")
+    return f"/{path[0]}" + path[2:].replace("\\", "/")
+
 
 def _rmtree_error_handler(func, path, exc_info):
     if not os.access(path, os.W_OK):
         # Is the error an access error ?
         os.chmod(path, stat.S_IWUSR)
         func(path)
-        print_debug('rmtree:read-only file/path (%s)' % (path, ))
+        log.debug(f"rmtree:read-only file/path ({path})")
     else:
         raise
 
+
 def rmtree_full(dest_dir, retry=False):
+    if not os.path.isdir(dest_dir):
+        return
     if retry:
-        for delay in [ 0.1, 0.2, 0.4, 0.8]:
+        for delay in [0.1, 0.2, 0.4, 0.8]:
             try:
                 shutil.rmtree(dest_dir, onerror=_rmtree_error_handler)
                 break
-            except:
+            except OSError:
                 # wait a little, don't ask me why ;(
                 time.sleep(delay)
     else:
         shutil.rmtree(dest_dir, onerror=_rmtree_error_handler)
 
+
 def read_file(file_name):
-    with open(file_name, 'rt') as fi:
-        rt = [ line.rstrip('\n') for line in fi ]
+    with open(file_name, encoding="utf-8") as fi:
+        rt = [line.rstrip("\n") for line in fi]
     return rt
 
+
 def write_file(file_name, content):
-    with open(file_name, 'wt') as fo:
+    with open(file_name, "w", encoding="utf-8") as fo:
         for i in content:
-            fo.write('%s\n' % (i, ))
+            fo.write(f"{i}\n")
+
 
 def file_replace(file_name, chg_list, make_bak=True):
-    """
-    Execute a series of replace on the file indicated
+    """Execute a series of replace on the file indicated.
 
     chg_list is an iterable of tuple (find, replace) to execute
     """
@@ -83,16 +87,17 @@ def file_replace(file_name, chg_list, make_bak=True):
         fc = nw
     if chg:
         if make_bak:
-            write_file(file_name + '.bak', sv)
+            write_file(f"{file_name}.bak", sv)
         write_file(file_name, fc)
+
 
 class ordered_set(set):
     def __init__(self):
         set.__init__(self)
-        self.__list = list()
+        self.__list = []
 
     def add(self, o):
-        if not o in self:
+        if o not in self:
             set.add(self, o)
             self.__list.append(o)
 
@@ -104,36 +109,29 @@ class ordered_set(set):
     def __iter__(self):
         return self.__list.__iter__()
 
-def python_find_libs_dir(org_dir):
-    """
-    From the python org_dir that can be also a virtualenv path
-    return the libs dir
-    """
 
-    cur = os.path.join(org_dir, 'libs')
+def python_find_libs_dir(org_dir):
+    """From the python org_dir that can be also a virtualenv path return the
+    libs dir."""
+
+    cur = os.path.join(org_dir, "libs")
     if os.path.isdir(cur):
         # easy :)
         return cur
 
     # look for the virtualenv marker
-    chk = os.path.join(org_dir, 'lib')
-    if not os.path.isdir(chk):
+    pyvenv_file = os.path.join(org_dir, "pyvenv.cfg")
+    if not os.path.isfile(pyvenv_file):
         # one level up
-        chk = os.path.join(org_dir, '..', 'lib')
+        pyvenv_file = os.path.join(org_dir, "..", "pyvenv.cfg")
 
-    if not chk:
-        # oops
-        return None
-
-    orig_file = os.path.join(chk, 'orig-prefix.txt')
-    if os.path.isfile(orig_file):
-        # Read and see whats happening
-        with open(orig_file, 'rt') as fi:
-            org_dir = fi.read()
-
+    if os.path.isfile(pyvenv_file):
+        # Read file to get home path
+        for line in read_file(pyvenv_file):
+            key, value = line.split("=")
+            if "home" == key.strip():
+                org_dir = value.strip()
+                break
     # Let's see if now is ok ..
-    cur = os.path.join(org_dir, 'libs')
-    if os.path.isdir(cur):
-        return cur
-
-    return None
+    cur = os.path.join(org_dir, "libs")
+    return cur if os.path.isdir(cur) else None
